final touch ups
