+++ src/App.js
@@ -1,14 +1,10 @@
 import React from 'react';
-// import logo from './logo.svg';
 import Home from './containers/home';
 import './App.css';
 
 function App() {
   return (
     <div className="App">
-      {/* <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-      </header> */}
       <Home />
     </div>
   );
