+++ src/elements/mainAvatar.jsx
@@ -18,10 +18,14 @@ export default function MainAvatar() {
   const classes = useStyles();
 
   return (
-    <>
+    <div>
       <Grid container justify="center">
         <Box pt={3}>
-          <Avatar alt="Random" src="https://randomuser.me/api/portraits/lego/0.jpg" className={classes.large} />
+          <Avatar
+            alt="Random" 
+            src="https://randomuser.me/api/portraits/lego/0.jpg"
+            className={classes.large}
+          />
         </Box>
       </Grid>
       <h2>
@@ -29,6 +33,6 @@ export default function MainAvatar() {
           The case of missing bicycles
         </p>
       </h2>
-    </>
+    </div>
   );
 }
