+++ src/reducers/reports.jsx
@@ -1,6 +1,4 @@
 import {
-  CREATE_REPORT,
-  REMOVE_REPORT,
   REPORT_BEGIN,
   REPORT_SUCCESS,
   REPORT_FAILURE,
@@ -11,10 +9,6 @@ const reportReducer = (
   action,
 ) => {
   switch (action.type) {
-    case CREATE_REPORT:
-      return [action.report];
-    case REMOVE_REPORT:
-      return state.filter(report => report.id !== action.report.id);
     case REPORT_BEGIN:
       return { loading: true, error: null };
     case REPORT_SUCCESS:
